requirements.txt
# BigQuery Integration
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.19.0

# Cloud Platform Integration
google-cloud-storage>=2.7.0
google-cloud-monitoring>=2.11.0
google-cloud-logging>=3.5.0

# Configuration Management
PyYAML>=6.0

# Date and Time Processing
python-dateutil>=2.8.2

# Reliability and Performance
backoff>=2.2.1
pyarrow>=12.0.0 
psutil>=5.9.0

# Development and Debugging
pylint>=2.17.0
black>=23.3.0
isort>=5.12.0